# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fresh Platform API
  description: |
    Complete API documentation for the Fresh Platform - a Georgian AI & Technology platform.
    
    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    Admin endpoints require a separate admin token via Cookie (`admin_session`).
    
    ## Rate Limiting
    - Search API: 30 requests per minute per IP
    - Admin login: Brute-force protection with lockout
    
    ## Georgian Language
    Many error messages and content are in Georgian.
    
  version: 1.0.0
  contact:
    name: Andrew Altair
    email: andrewaltair@icloud.com
  license:
    name: MIT
servers:
  - url: https://api.example.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: User login
      description: Login with email/username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                email:
                  type: string
                username:
                  type: string
                password:
                  type: string
            example:
              email: "user@example.com"
              password: "securePassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
        '403':
          description: Account blocked
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      description: Get authenticated user data
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User data
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
...
  /api/posts/generate-tags:
    post:
      tags: [Posts]
      summary: Generate tags with AI
      description: Generate 20 SEO-optimized tags using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                excerpt:
                  type: string
                content:
                  type: string
                category:
                  type: string
      responses:
        '200':
          description: Generated tags
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  tags:
                    type: array
                    items:
                      type: string
...
  /api/mystic/fortune:
    post:
      tags: [Mystic]
      summary: Fortune telling
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                birthDate:
                  type: string
      responses:
        '200':
          description: Fortune
          content:
            application/json:
              schema:
                type: object
                properties:
                  prediction:
                    type: string
                  luckyColor:
                    type: string
                  luckyNumber:
                    type: string
                  luckyDay:
                    type: string
...
  /api/email:
    post:
      tags: [Email]
      summary: Send email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, to]
              properties:
                type:
                  type: string
                  enum: [welcome, notification, custom]
                to:
                  type: string
                  format: email
                name:
                  type: string
                  description: For welcome email
                title:
                  type: string
                  description: For notification email
                message:
                  type: string
                  description: For notification email
                actionUrl:
                  type: string
                actionText:
                  type: string
                subject:
                  type: string
                  description: For custom email
                html:
                  type: string
                  description: For custom email
      responses:
        '200':
          description: Email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  messageId:
                    type: string
                  previewUrl:
                    type: string

  # ============ SEO ============
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminCookie:
      type: apiKey
      in: cookie
      name: admin_session
  schemas:
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        avatar:
          type: string
        coverImage:
          type: string
        role:
          type: string
          enum: [viewer, member, author, admin, god]
        badge:
          type: string
        createdAt:
          type: string
          format: date-time
...